I #1525: Separate reduction-related built-in tests
