+++ src/psyclone/tests/domain/lfric/lfric_builtins_test.py
@@ -3643,242 +3643,6 @@ def test_setval_X(tmpdir, monkeypatch, annexed, dist_mem, fortran_writer):
             output_dm_2 = output_dm_2.replace("dof_annexed", "dof_owned")
         assert output_dm_2 in code
 
-# ------------- Inner product of real fields -------------------------------- #
-
-
-def test_X_innerproduct_Y(tmpdir, dist_mem):
-    '''Test that 1) the '__str__' method of 'LFRicXInnerproductYKern'
-    returns the expected string and 2) we generate correct code for
-    the built-in operation which calculates inner product of
-    real-valued fields 'X' and 'Y' as 'innprod = innprod +
-    X(:)*Y(:)'. 3) Also test the 'metadata()' method.
-
-    '''
-    metadata = lfric_builtins.LFRicXInnerproductYKern.metadata()
-    assert isinstance(metadata, LFRicKernelMetadata)
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH,
-                     "15.9.1_X_innerproduct_Y_builtin.f90"),
-        api=API)
-    psy = PSyFactory(API,
-                     distributed_memory=dist_mem).create(invoke_info)
-    # Test '__str__' method
-    first_invoke = psy.invokes.invoke_list[0]
-    kern = first_invoke.schedule.children[0].loop_body[0]
-    assert str(kern) == "Built-in: X_innerproduct_Y (real-valued fields)"
-    # Test code generation
-    code = str(psy.gen)
-
-    assert LFRicBuild(tmpdir).code_compiles(psy)
-
-    output = (
-        "      !\n"
-        "      ! Initialise field and/or operator proxies\n"
-        "      !\n"
-        "      f1_proxy = f1%get_proxy()\n"
-        "      f1_data => f1_proxy%data\n"
-        "      f2_proxy = f2%get_proxy()\n"
-        "      f2_data => f2_proxy%data\n"
-        "      !\n")
-    assert output in code
-    if not dist_mem:
-        output_seq = (
-            "      ! Initialise number of DoFs for aspc1_f1\n"
-            "      !\n"
-            "      undf_aspc1_f1 = f1_proxy%vspace%get_undf()\n"
-            "      !\n"
-            "      ! Set-up all of the loop bounds\n"
-            "      !\n"
-            "      loop0_start = 1\n"
-            "      loop0_stop = undf_aspc1_f1\n"
-            "      !\n"
-            "      ! Call our kernels\n"
-            "      !\n"
-            "      !\n"
-            "      ! Zero summation variables\n"
-            "      !\n"
-            "      asum = 0.0_r_def\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        asum = asum + f1_data(df)*f2_data(df)\n"
-            "      END DO\n"
-            "      !\n")
-        assert output_seq in code
-    else:
-        output_dm = (
-            "      loop0_stop = f1_proxy%vspace%get_last_dof_owned()\n"
-            "      !\n"
-            "      ! Call kernels and communication routines\n"
-            "      !\n"
-            "      !\n"
-            "      ! Zero summation variables\n"
-            "      !\n"
-            "      asum = 0.0_r_def\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        asum = asum + f1_data(df)*f2_data(df)\n"
-            "      END DO\n"
-            "      global_sum%value = asum\n"
-            "      asum = global_sum%get_sum()\n"
-            "      !\n")
-        assert output_dm in code
-        assert "      USE scalar_mod, ONLY: scalar_type" in code
-        assert "      REAL(KIND=r_def), intent(out) :: asum\n" in code
-        assert "      TYPE(scalar_type) global_sum\n" in code
-
-
-def test_X_innerproduct_X(tmpdir, dist_mem):
-    '''Test that 1) the '__str__' method of 'LFRicXInnerproductXKern'
-    returns the expected string and 2) we generate correct code for
-    the built-in operation which calculates inner product of a
-    real-valued field 'X' by itself as 'innprod = innprod +
-    X(:)*X(:)'. 3) Also test the 'metadata()' method.
-
-    '''
-    metadata = lfric_builtins.LFRicXInnerproductXKern.metadata()
-    assert isinstance(metadata, LFRicKernelMetadata)
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH,
-                     "15.9.2_X_innerproduct_X_builtin.f90"),
-        api=API)
-    psy = PSyFactory(API,
-                     distributed_memory=dist_mem).create(invoke_info)
-    # Test '__str__' method
-    first_invoke = psy.invokes.invoke_list[0]
-    kern = first_invoke.schedule.children[0].loop_body[0]
-    assert str(kern) == "Built-in: X_innerproduct_X (real-valued field)"
-    # Test code generation
-    code = str(psy.gen)
-
-    assert LFRicBuild(tmpdir).code_compiles(psy)
-
-    output = (
-        "      !\n"
-        "      ! Initialise field and/or operator proxies\n"
-        "      !\n"
-        "      f1_proxy = f1%get_proxy()\n"
-        "      f1_data => f1_proxy%data\n"
-        "      !\n")
-    assert output in code
-    if not dist_mem:
-        output_seq = (
-            "      ! Initialise number of DoFs for aspc1_f1\n"
-            "      !\n"
-            "      undf_aspc1_f1 = f1_proxy%vspace%get_undf()\n"
-            "      !\n"
-            "      ! Set-up all of the loop bounds\n"
-            "      !\n"
-            "      loop0_start = 1\n"
-            "      loop0_stop = undf_aspc1_f1\n"
-            "      !\n"
-            "      ! Call our kernels\n"
-            "      !\n"
-            "      !\n"
-            "      ! Zero summation variables\n"
-            "      !\n"
-            "      asum = 0.0_r_def\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        asum = asum + f1_data(df)*f1_data(df)\n"
-            "      END DO\n"
-            "      !\n")
-        assert output_seq in code
-    else:
-        output_dm = (
-            "      loop0_stop = f1_proxy%vspace%get_last_dof_owned()\n"
-            "      !\n"
-            "      ! Call kernels and communication routines\n"
-            "      !\n"
-            "      !\n"
-            "      ! Zero summation variables\n"
-            "      !\n"
-            "      asum = 0.0_r_def\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        asum = asum + f1_data(df)*f1_data(df)\n"
-            "      END DO\n"
-            "      global_sum%value = asum\n"
-            "      asum = global_sum%get_sum()\n"
-            "      !\n")
-        assert output_dm in code
-        assert "      USE scalar_mod, ONLY: scalar_type" in code
-        assert "      REAL(KIND=r_def), intent(out) :: asum\n" in code
-        assert "      TYPE(scalar_type) global_sum\n" in code
-
-
-# ------------- Sum real field elements ------------------------------------- #
-
-
-def test_sum_X(tmpdir, dist_mem):
-    '''Test that 1) the '__str__' method of 'LFRicSumXKern' returns the
-    expected string and 2) we generate correct code for the built-in
-    operation which sums elements of a real-valued field 'X' as
-    'sumfld = sum(X(:))'. 3) Also test the 'metadata()' method.
-
-    '''
-    metadata = lfric_builtins.LFRicSumXKern.metadata()
-    assert isinstance(metadata, LFRicKernelMetadata)
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH,
-                     "15.8.1_sum_X_builtin.f90"), api=API)
-    psy = PSyFactory(API,
-                     distributed_memory=dist_mem).create(invoke_info)
-    # Test '__str__' method
-    first_invoke = psy.invokes.invoke_list[0]
-    kern = first_invoke.schedule.children[0].loop_body[0]
-    assert str(kern) == "Built-in: sum_X (sum a real-valued field)"
-    # Test code generation
-    code = str(psy.gen)
-
-    assert LFRicBuild(tmpdir).code_compiles(psy)
-
-    output = (
-        "      !\n"
-        "      ! Initialise field and/or operator proxies\n"
-        "      !\n"
-        "      f1_proxy = f1%get_proxy()\n"
-        "      f1_data => f1_proxy%data\n"
-        "      !\n")
-    assert output in code
-    if not dist_mem:
-        output = (
-            "      undf_aspc1_f1 = f1_proxy%vspace%get_undf()\n"
-            "      !\n"
-            "      ! Set-up all of the loop bounds\n"
-            "      !\n"
-            "      loop0_start = 1\n"
-            "      loop0_stop = undf_aspc1_f1\n"
-            "      !\n"
-            "      ! Call our kernels\n"
-            "      !\n"
-            "      !\n"
-            "      ! Zero summation variables\n"
-            "      !\n"
-            "      asum = 0.0_r_def\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        asum = asum + f1_data(df)\n"
-            "      END DO")
-        assert output in code
-    else:
-        output = (
-            "      loop0_stop = f1_proxy%vspace%get_last_dof_owned()\n"
-            "      !\n"
-            "      ! Call kernels and communication routines\n"
-            "      !\n"
-            "      !\n"
-            "      ! Zero summation variables\n"
-            "      !\n"
-            "      asum = 0.0_r_def\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        asum = asum + f1_data(df)\n"
-            "      END DO\n"
-            "      global_sum%value = asum\n"
-            "      asum = global_sum%get_sum()")
-        assert output in code
-        assert "      REAL(KIND=r_def), intent(out) :: asum\n" in code
-
 
 # ------------- Sign of real field elements --------------------------------- #
 
@@ -4955,152 +4719,6 @@ def test_builtin_set_plus_normal(tmpdir, monkeypatch, annexed, dist_mem):
                 "      ! Call kernels and communication routines\n"
                 "      !\n", f1_hex_code)
         assert output_dm_2 in code
-
-
-# ------------- Built-ins with reductions ----------------------------------- #
-
-
-def test_multi_builtin_single_invoke(tmpdir, monkeypatch, annexed, dist_mem):
-    '''Test that multiple built-ins, including one with reductions,
-    produce correct code. Also test with and without annexed DoFs
-    being computed as this affects the generated code.
-
-    '''
-    api_config = Config.get().api_conf(API)
-    monkeypatch.setattr(api_config, "_compute_annexed_dofs", annexed)
-    _, invoke_info = parse(
-        os.path.join(BASE_PATH,
-                     "15.18.1_builtins_reduction_fuse_error.f90"),
-        api=API)
-    psy = PSyFactory(API, distributed_memory=dist_mem).create(invoke_info)
-    code = str(psy.gen)
-
-    assert LFRicBuild(tmpdir).code_compiles(psy)
-
-    if dist_mem:
-        assert (
-            "    SUBROUTINE invoke_0(asum, f1, f2, b)\n"
-            "      USE scalar_mod, ONLY: scalar_type\n"
-            "      USE mesh_mod, ONLY: mesh_type\n"
-            "      REAL(KIND=r_def), intent(out) :: asum\n"
-            "      REAL(KIND=r_def), intent(in) :: b\n"
-            "      TYPE(field_type), intent(in) :: f1, f2\n"
-            "      TYPE(scalar_type) global_sum\n"
-            "      INTEGER df\n"
-            "      INTEGER(KIND=i_def) loop2_start, loop2_stop\n"
-            "      INTEGER(KIND=i_def) loop1_start, loop1_stop\n"
-            "      INTEGER(KIND=i_def) loop0_start, loop0_stop\n"
-            "      REAL(KIND=r_def), pointer, dimension(:) :: f2_data => "
-            "null()\n"
-            "      REAL(KIND=r_def), pointer, dimension(:) :: f1_data => "
-            "null()\n"
-            "      TYPE(field_proxy_type) f1_proxy, f2_proxy\n") in code
-        assert (
-            "      f1_proxy = f1%get_proxy()\n"
-            "      f1_data => f1_proxy%data\n"
-            "      f2_proxy = f2%get_proxy()\n"
-            "      f2_data => f2_proxy%data\n"
-            "      !\n") in code
-        output = (
-            "      loop2_stop = f1_proxy%vspace%get_last_dof_annexed()\n"
-            "      !\n"
-            "      ! Call kernels and communication routines\n"
-            "      !\n"
-            "      !\n"
-            "      ! Zero summation variables\n"
-            "      !\n"
-            "      asum = 0.0_r_def\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        asum = asum + f1_data(df)*f2_data(df)\n"
-            "      END DO\n"
-            "      global_sum%value = asum\n"
-            "      asum = global_sum%get_sum()\n"
-            "      DO df=loop1_start,loop1_stop\n"
-            "        f1_data(df) = b * f1_data(df)\n"
-            "      END DO\n"
-            "      !\n"
-            "      ! Set halos dirty/clean for fields modified in the "
-            "above loop\n"
-            "      !\n"
-            "      CALL f1_proxy%set_dirty()\n"
-            "      !\n"
-            "      DO df=loop2_start,loop2_stop\n"
-            "        f1_data(df) = asum * f1_data(df)\n"
-            "      END DO\n"
-            "      !\n"
-            "      ! Set halos dirty/clean for fields modified in the "
-            "above loop\n"
-            "      !\n"
-            "      CALL f1_proxy%set_dirty()\n")
-        if not annexed:
-            output = output.replace("dof_annexed", "dof_owned")
-        assert output in code
-    else:
-        assert (
-            "    SUBROUTINE invoke_0(asum, f1, f2, b)\n"
-            "      REAL(KIND=r_def), intent(out) :: asum\n"
-            "      REAL(KIND=r_def), intent(in) :: b\n"
-            "      TYPE(field_type), intent(in) :: f1, f2\n"
-            "      INTEGER df\n"
-            "      INTEGER(KIND=i_def) loop2_start, loop2_stop\n"
-            "      INTEGER(KIND=i_def) loop1_start, loop1_stop\n"
-            "      INTEGER(KIND=i_def) loop0_start, loop0_stop\n"
-            "      REAL(KIND=r_def), pointer, dimension(:) :: f2_data => "
-            "null()\n"
-            "      REAL(KIND=r_def), pointer, dimension(:) :: f1_data => "
-            "null()\n"
-            "      TYPE(field_proxy_type) f1_proxy, f2_proxy\n"
-            "      INTEGER(KIND=i_def) undf_aspc1_f1\n") in code
-        assert (
-            "      f1_proxy = f1%get_proxy()\n"
-            "      f1_data => f1_proxy%data\n"
-            "      f2_proxy = f2%get_proxy()\n"
-            "      f2_data => f2_proxy%data\n"
-            "      !\n"
-            "      ! Initialise number of DoFs for aspc1_f1\n"
-            "      !\n"
-            "      undf_aspc1_f1 = "
-            "f1_proxy%vspace%get_undf()\n" in code)
-        assert (
-            "      loop0_stop = undf_aspc1_f1\n"
-            "      loop1_start = 1\n"
-            "      loop1_stop = undf_aspc1_f1\n"
-            "      loop2_start = 1\n"
-            "      loop2_stop = undf_aspc1_f1\n" in code)
-        assert (
-            "      asum = 0.0_r_def\n"
-            "      !\n"
-            "      DO df=loop0_start,loop0_stop\n"
-            "        asum = asum + f1_data(df)*f2_data(df)\n"
-            "      END DO\n"
-            "      DO df=loop1_start,loop1_stop\n"
-            "        f1_data(df) = b * f1_data(df)\n"
-            "      END DO\n"
-            "      DO df=loop2_start,loop2_stop\n"
-            "        f1_data(df) = asum * f1_data(df)\n"
-            "      END DO\n") in code
-
-
-# ------------- Invalid built-in with an integer scalar reduction ----------- #
-
-
-def test_scalar_int_builtin_error(monkeypatch):
-    '''Test that specifying incorrect meta-data for built-in such that it
-    claims to perform a reduction into an integer variable raises the
-    expected error.
-
-    '''
-    monkeypatch.setattr(lfric_builtins, "BUILTIN_DEFINITIONS_FILE",
-                        value=os.path.join(BASE_PATH,
-                                           "int_reduction_builtins_mod.f90"))
-    with pytest.raises(ParseError) as excinfo:
-        _, _ = parse(os.path.join(BASE_PATH,
-                                  "16.2_integer_scalar_sum.f90"),
-                     api=API)
-    assert ("In the LFRic API a reduction access 'gh_sum' is only valid "
-            "with a real scalar argument, but a scalar argument with "
-            "'gh_integer' data type was found" in str(excinfo.value))
 
 
 # ------------- Auxiliary mesh code generation function --------------------- #
